co-biz-anlyzer/requirements.txt
crewai>=0.51.0
crewai-tools>=0.4.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
